Extract project-related helpers to their own file
